HBX-1658: Move interface 'org.hibernate.tool.hbm2x.pojo.POJOClass' to 'org.hibernate.tool.internal.export.pojo.POJOClass'
