+++ orm/src/main/java/org/hibernate/tool/hbm2x/GenericExporter.java
@@ -11,9 +11,9 @@ import java.util.StringTokenizer;
 import org.hibernate.internal.util.StringHelper;
 import org.hibernate.mapping.Component;
 import org.hibernate.tool.hbm2x.pojo.ComponentPOJOClass;
-import org.hibernate.tool.hbm2x.pojo.POJOClass;
 import org.hibernate.tool.internal.export.common.AbstractExporter;
 import org.hibernate.tool.internal.export.common.ConfigurationNavigator;
+import org.hibernate.tool.internal.export.pojo.POJOClass;
 
 
 public class GenericExporter extends AbstractExporter {
