+++ orm/src/main/java/org/hibernate/tool/hbm2x/pojo/ComponentPOJOClass.java
@@ -9,6 +9,7 @@ import java.util.List;
 import org.hibernate.mapping.Component;
 import org.hibernate.mapping.Property;
 import org.hibernate.tool.internal.export.pojo.Cfg2JavaTool;
+import org.hibernate.tool.internal.export.pojo.POJOClass;
 
 public class ComponentPOJOClass extends BasicPOJOClass {
