+++ orm/src/main/java/org/hibernate/tool/internal/export/dao/DAOExporter.java
@@ -2,7 +2,7 @@ package org.hibernate.tool.internal.export.dao;
 
 import java.util.Map;
 
-import org.hibernate.tool.hbm2x.pojo.POJOClass;
+import org.hibernate.tool.internal.export.pojo.POJOClass;
 import org.hibernate.tool.internal.export.pojo.POJOExporter;
 
 public class DAOExporter extends POJOExporter {
